Fix infinite loop for shell sourcing on RHEL systems
Prior to this commit, any non-root user when running `bash` on the cli would get a segfault. This happens because `/etc/bashrc` sources `/etc/profile.d/*.sh` and this file in there source `/etc/bashrc` and you can see we've now created a loop. This just checks if you're on RH and does nothing if that's the case since `/etc/bashrc` will have already been sourced.